Addresses some issues Functional testing comments #3
PrestaShop 1.6 -> Confirmation message not properly shown (no design
applied)
*** Design Applied

PrestaShop 1.6 -> (Improvement) Use bootstrap theme on the settings page
*** Bootstrap implemented

PrestaShop 1.6 -> Block added on order details is not properly shown
and should be improved
*** Block UI improved

You should not have two versions of the configuration page. The Mexican
version is enough
*** non-mexican version removed

PrestaShop 1.4 -> Design is broken when the module is enabled. The
string which contains CSS definition must be returned, not shown with
"echo"
*** Code changed for 1.4 ->

PrestaShop 1.6 -> Express checkout button is not on a good place. It is
under another logos
*** UI for button edited